+++ NEWS.md
@@ -1,4 +1,11 @@
 
+# NCovUtils 0.2.0
+
+* Added `memoise` functionality to automatically cache in the current directory all remote resources.
+* Added a non-Hubei linelist function that can extract country and city linelists.
+* Added a function to extract the latest WHO case counts.
+* Added a function to summarise the current total number of cases.
+
 # NCoVUtils 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.